Safari-611.1.10/LayoutTests/crypto/workers/subtle/ec-import-jwk-key-public-expected.txt

[Worker] Test importing a JWK ECDH public key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-import-jwk-key-public.js
[Worker] Importing a key...
PASS [Worker] publicKey.toString() is '[object CryptoKey]'
PASS [Worker] publicKey.type is 'public'
PASS [Worker] publicKey.extractable is true
PASS [Worker] publicKey.algorithm.name is 'ECDH'
PASS [Worker] publicKey.algorithm.namedCurve is 'P-384'
PASS [Worker] publicKey.usages is [ ]
PASS successfullyParsed is true

TEST COMPLETE
